requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
beautifulsoup4==4.11.1
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
django==4.1.4
django-debug-toolbar==3.8.1
django-environ==0.9.0
django-timezone-field==5.0
djangorestframework==3.14.0
drf-spectacular==0.25.1
drf-ujson2==1.7.2
idna==3.4
inflection==0.5.1
jsonschema==4.17.3
kombu==5.2.4
prompt-toolkit==3.0.36
psycopg2==2.9.5
pydantic==1.10.4
pyrsistent==0.19.2
pytz==2022.7
pyyaml==6.0
redis==4.4.2
requests==2.28.1
requests-futures==1.0.0
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
typing-extensions==4.4.0
tzdata==2022.7
ujson==5.7.0
uritemplate==4.1.1
urllib3==1.26.13
uwsgi==2.0.21
vine==5.0.0
wcwidth==0.2.6
whitenoise==6.2.0
